src/requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
freetype-py==2.2.0
mypy-extensions==0.4.3
numpy==1.19.5
pathspec==0.8.1
PyOpenGL==3.1.5
PyQt5==5.15.2
PyQt5-sip==12.8.1
regex==2020.11.13
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
